+++ src/locales/en/index.ts
@@ -21,7 +21,6 @@ import lorem from './lorem';
 import music from './music';
 import person from './person';
 import phone_number from './phone_number';
-import science from './science';
 import team from './team';
 import vehicle from './vehicle';
 import word from './word';
@@ -46,7 +45,6 @@ const en: LocaleDefinition = {
   music,
   person,
   phone_number,
-  science,
   team,
   vehicle,
   word,
