+++ src/locales/global/finance/currency.ts
