+++ src/locales/global/science/chemicalElement.ts
