+++ src/locales/global/science/index.ts
